requirements.txt
requests
requests-oauthlib
python-dotenv
beautifulsoup4
fpdf2